% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as_markdown}
\alias{as_markdown}
\title{Convert to markdown chunk}
\usage{
as_markdown(..., collapse = "")
}
\arguments{
\item{...}{dots to convert to character vector}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.  When \code{collapse} is a string,
    the result is always a string (\code{\link[base]{character}} of length 1).}
}
\value{
character vector of length 1
}
\description{
This is a wrapper around \link[base:paste]{base::paste0} which sets the class of the resulting
character vector to "knit_asis". This means that the \link[knitr:knitr-package]{knitr::knitr} engine will render
the character vector as markdown content, rather than as an R value.
}
\examples{
as_markdown(c("Hello world!"))
}
\keyword{internal}
