% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{check_src}
\alias{check_src}
\title{Check source string}
\usage{
check_src(
  x,
  allow_missing = FALSE,
  allow_empty = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{Input string. Typically a file path or URL.}

\item{allow_missing}{If \code{FALSE} (default), error if string is not a URL and
not a valid web path. If \code{TRUE}, allow input to be a path to a non-existent
file.}

\item{allow_empty}{If \code{FALSE} (default), error if string is empty. If \code{TRUE},
allow input to be an empty string.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Used by \code{\link[=qto_src_span]{qto_src_span()}} to check input strings.
}
\keyword{internal}
