% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine}
\alias{combine}
\title{Combine a string with text before and after}
\usage{
combine(x, before = "", after = before, allow_empty = TRUE, allow_null = TRUE)
}
\arguments{
\item{x}{An input string.}

\item{before, after}{Text to insert before and after the supplied string.}

\item{allow_empty}{If \code{TRUE}, allow empty string ("") values as input. If
\code{FALSE}, error on empty string inputs.}

\item{allow_null}{If \code{TRUE}, allow \code{NULL} values as input (replaced with an
empty string ""). If \code{FALSE}, error on \code{NULL} inputs.}
}
\description{
Combine a string with text before and after
}
\keyword{internal}
