% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{div}
\alias{div}
\title{Create a quarto div element}
\usage{
div(..., attr = NULL)
}
\arguments{
\item{...}{todoc}

\item{attr}{todoc}
}
\value{
character vector of length 1
}
\description{
Create a div block element with optional attributes to assign to it. Dots
is collapsed into a single character vector with a newline between each element.

If attr is a vector with a length > 1,
attributes will be combined into a single character vector via \link[base:paste]{base::paste0}.
}
\examples{
div("Hello world!", attr = ".callout-tip")
}
\seealso{
Other elements: 
\code{\link{span}()}
}
\concept{elements}
