% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-list.R
\name{list_drop_or_replace_na}
\alias{list_drop_or_replace_na}
\title{Drop or replace NA items from a list}
\usage{
list_drop_or_replace_na(x, drop_na = FALSE, replacement = "")
}
\arguments{
\item{drop_na}{If \code{TRUE}, drop \code{NA} values from the supplied list.}

\item{replacement}{If not \code{NULL}, replace NA values from the supplied list
with the value of replacement.}
}
\description{
Drop or replace NA items from a list
}
\keyword{internal}
