% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_qto}
\alias{map_qto}
\title{Apply a function to each element of a vector and return Quarto block vector}
\usage{
map_qto(.x, .f = NULL, ..., .type = c("block", "div", "callout", "heading"))
}
\arguments{
\item{.x}{An input vector.}

\item{.f}{Optional function to apply to each element. If function does not
return a "quarto_block" class object, the output is passed to \code{\link[=qto_block]{qto_block()}}}

\item{.type}{If .f is \code{NULL}, type is used to define the function applied to
each element of the vector. Options include "block", "div", "callout", or
"heading".}
}
\description{
Apply a function to each element of a vector and return Quarto block vector
}
\examples{
map_qto(
  list("This is a note.", "And this is a note.", "And this is a note"),
  .type = "callout"
)

}
