% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mdapply}
\alias{mdapply}
\title{Markdown Apply}
\usage{
mdapply(x, fun, ...)
}
\arguments{
\item{x}{a vector, expression, or any object coercible to a list}

\item{fun}{function to apply to each element of x}

\item{...}{optional arguments to fun}
}
\value{
character vector of length 1
}
\description{
Generate markdown by applying a function to a list of elements.
The function is expected to return an object that can be coerced to a character
vector.

The final result will be converted via \link{as_markdown} to a valid
markdown chunk.
}
\examples{
my_list <- list("A", "B", "C")
mdapply(my_list, div)
}
\seealso{
\link[base:lapply]{base::lapply} \link{as_markdown}
}
\keyword{internal}
