% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{pmap_qto}
\alias{pmap_qto}
\title{Map over multiple inputs simultaenously and return Quarto block vector}
\usage{
pmap_qto(
  .l,
  .f = NULL,
  ...,
  .type = c("block", "div", "callout", "heading"),
  .sep = "",
  .collapse = "",
  call = caller_env()
)
}
\arguments{
\item{.l}{An input vector.}

\item{.f}{Optional function to apply to each element. If function does not
return a "quarto_block" class object, the output is passed to \code{\link[=qto_block]{qto_block()}}}

\item{...}{Additional parameters passed to function defined by \code{.f}.}

\item{.type}{If .f is \code{NULL}, type is used to define the function applied to
each element of the vector. Options include "block", "div", "callout", or
"heading".}

\item{.sep, .collapse}{Additional parameters passed to \code{\link[=qto_block]{qto_block()}} if .f
does not return a quarto block class object. Ignored if .f does return a
quarto block class object.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=pmap_qto]{pmap_qto()}} loops a list of vectors over a package function defined by .type or a custom
function that returns a quarto block output. This function always returns a
list of quarto block objects.
}
\examples{
qto_list <- pmap_qto(
    list(
        list("Answer:", "Answer:", "Answer:"),
        list("Yes", "No", "Yes")
    )
)
qto_block(qto_list)

qto_list <- pmap_qto(
    mtcars[seq(3L), seq(3L)],
    function(mpg, cyl, disp) {
        qto_li(
            .list = list(
                sprintf("mpg is: \%s", mpg),
                sprintf("cyl is: \%s", cyl),
                sprintf("disp is: \%s", disp)
            )
        )
    }
)
qto_block(qto_list)

}
\seealso{
\code{\link[=map_qto]{map_qto()}}, \code{\link[purrr:pmap]{purrr::pmap()}}
}
