% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{qto_basics}
\alias{qto_basics}
\alias{qto_heading}
\alias{qto_hr}
\title{Basic formatting and markdown elements}
\usage{
qto_heading(..., level = 1L)

qto_hr(rule = "-", length = 72L, before = "\\n\\n", after = before)
}
\arguments{
\item{...}{Heading text passed sto \code{\link[=qto_block]{qto_block()}}.}

\item{level}{Heading level. Defaults to 1.}

\item{rule, length}{Horizontal rule character and length of rule.}

\item{before, after}{Text to insert before and after a horizontal rule.}
}
\description{
\code{\link[=qto_heading]{qto_heading()}} creates headings and \code{\link[=qto_hr]{qto_hr()}} creates horizontal rules.
}
