% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callout.R
\name{qto_callout}
\alias{qto_callout}
\title{Create a Quarto callout block}
\usage{
qto_callout(
  ...,
  type = c("note", "tip", "warning", "caution", "important"),
  collapse = NULL,
  appearance = NULL,
  icon = NULL,
  title = NULL,
  id = NULL,
  class = NULL,
  .attributes = NULL,
  .call = caller_env()
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=qto_div]{qto_div}}
  \describe{
    \item{\code{.content}}{If \code{.content} is supplied, any values passed to \code{...} are
ignored. If \code{.content} is \code{NULL}, it is set as all values passed to \code{...}.}
    \item{\code{drop_empty}}{If \code{TRUE}, drop empty values from \code{.content} or \code{...}}
    \item{\code{drop_na}}{If \code{TRUE}, drop \code{NA} values from \code{.content} or \code{...}}
    \item{\code{css}}{If \code{{htmltools}} is installed, a list of css style attributes to
pass to \code{\link[htmltools:css]{htmltools::css()}}.}
  }}

\item{type}{Callout type. One of "note", "tip", "warning", "caution", or
"important". \url{https://quarto.org/docs/authoring/callouts.html#callout-types}}

\item{collapse}{If \code{TRUE}, create a folded callout. \url{https://quarto.org/docs/authoring/callouts.html#collapse}}

\item{appearance}{Callout appearance. One of "default", "simple", "minimal".
\url{https://quarto.org/docs/authoring/callouts.html#appearance}}

\item{icon}{If \code{FALSE}, create a callout with an icon. \url{https://quarto.org/docs/authoring/callouts.html#icons}}

\item{title}{Title attribute. Works the same as included a level 2 heading
before the body of the callout contents.}

\item{id}{Div or span identifier. If \code{id} does not start with \code{"#"}, the
hash character is applied as a prefix.}

\item{class}{Div or span class. If \code{class} does not start with \code{"."}, the
period character is applied as a prefix.}

\item{.attributes}{Optional list of attributes. If supplied, any attributes
passed to \code{...} are ignored.}

\item{.call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Callouts are an excellent way to draw extra attention to certain concepts, or
to more clearly indicate that certain content is supplemental or applicable
to only some scenarios. \url{https://quarto.org/docs/authoring/callouts.html}
}
\examples{

qto_callout(
  "Callouts provide a simple way to attract attention, for example, to this warning.",
  type = "warning"
)

qto_callout(
  "This is an example of a callout with a title.",
  type = "tip",
  title = "Tip with a title"
)

}
