% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{qto_definition_list}
\alias{qto_definition_list}
\title{Create a definition list}
\usage{
qto_definition_list(
  ...,
  .term = NULL,
  .definition = NULL,
  .list = NULL,
  .drop_na = FALSE,
  .sep = "\\n",
  .replacement = ""
)
}
\arguments{
\item{...}{Optional. Named argument where the argument name is a term and the
value name is the text to display as a definition for the term. Ignored if
\code{.list} or \code{.term} and \code{.definition} are supplied.}

\item{.term, .definition}{If supplied, \code{.list} is set to a named list using
terms as names and definitions as values. Ignored if \code{.list} is supplied.}

\item{.list}{If \code{.list} is supplied, any values passed to \code{...} are ignored.
One of \code{.list}, \code{.term} and \code{.definition}, or \code{...} must be supplied.}

\item{.drop_na}{If \code{TRUE}, drop \code{NA} values from the supplied list.}

\item{.replacement}{If not \code{NULL}, replace NA values from the supplied list
with the value of replacement.}
}
\description{
Create a definition list
}
