% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{qto_div}
\alias{qto_div}
\title{Create a Quarto div with optional classes, attributes, and other identifiers}
\usage{
qto_div(
  ...,
  id = NULL,
  class = NULL,
  css = NULL,
  .attributes = NULL,
  .content = NULL,
  collapse = "",
  drop_empty = TRUE,
  drop_na = TRUE,
  call = caller_env()
)
}
\arguments{
\item{...}{Optional named attributes.}

\item{id}{Div or span identifier. If \code{id} does not start with \code{"#"}, the
hash character is applied as a prefix.}

\item{class}{Div or span class. If \code{class} does not start with \code{"."}, the
period character is applied as a prefix.}

\item{css}{If \code{{htmltools}} is installed, a list of css style attributes to
pass to \code{\link[htmltools:css]{htmltools::css()}}.}

\item{.attributes}{Optional list of attributes. If supplied, any attributes
passed to \code{...} are ignored.}

\item{.content}{If \code{.content} is supplied, any values passed to \code{...} are
ignored. If \code{.content} is \code{NULL}, it is set as all values passed to \code{...}.}

\item{collapse}{Passed to \code{\link[base:paste]{base::paste0()}} with \code{.content.}}

\item{drop_empty}{If \code{TRUE}, drop empty values from \code{.content} or \code{...}}

\item{drop_na}{If \code{TRUE}, drop \code{NA} values from \code{.content} or \code{...}}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\url{https://quarto.org/docs/authoring/markdown-basics.html#divs-and-spans}
}
\examples{

# div with an class
qto_div(
  "This content can be styled with a border",
  class = "border"
)

# Nested div
qto_div(
  qto_div("Here is a warning.", class = "warning"),
  "More content.",
  id = "special",
  class = "sidebar"
)

}
\seealso{
\itemize{
\item \code{\link[=qto_callout]{qto_callout()}}
}
}
