% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{qto_dl}
\alias{qto_dl}
\title{Create a definition list}
\usage{
qto_dl(
  ...,
  .term = NULL,
  .definition = NULL,
  .list = NULL,
  .drop_na = FALSE,
  .sep = "\\n",
  .replacement = ""
)
}
\arguments{
\item{...}{Optional. Named argument where the argument name is a term and the
value name is the text to display as a definition for the term. Ignored if
\code{.list} or \code{.term} and \code{.definition} are supplied.}

\item{.term, .definition}{If supplied, \code{.list} is set to a named list using
terms as names and definitions as values. Ignored if \code{.list} is supplied.}

\item{.list}{If \code{.list} is supplied, any values passed to \code{...} are ignored.
One of \code{.list}, \code{.term} and \code{.definition}, or \code{...} must be supplied.}

\item{.drop_na}{If \code{TRUE}, drop \code{NA} values from the supplied list.}

\item{.sep}{Added separator string between terms and definitions. Defaults to
\code{"\\n"} for wider list spacing. If set to \code{""}, the definition list uses a
tighter spacing.}

\item{.replacement}{If not \code{NULL}, replace NA values from the supplied list
with the value of replacement.}
}
\description{
See the Quarto documentation for more information on lists:
\url{https://quarto.org/docs/authoring/markdown-basics.html#lists}
}
\examples{
qto_dl(
  "Quarto" = "An open-source scientific and technical publishing system"
)

qto_dl(
  .term = "Quarto",
  .definition = "An open-source scientific and technical publishing system"
)

qto_dl(
  .list = list("Quarto" = "An open-source scientific and technical publishing system")
)

}
\seealso{
\itemize{
\item \code{\link[=qto_ol]{qto_ol()}}
\item \code{\link[=qto_li]{qto_li()}}
}
}
