% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{qto_fig_span}
\alias{qto_fig_span}
\title{Simplified helper for qto_fig()}
\usage{
qto_fig_span(
  src,
  caption = NULL,
  ...,
  allow_missing = TRUE,
  allow_empty = FALSE,
  call = caller_env()
)
}
\arguments{
\item{src}{Image source. Either a file path or URL.}

\item{caption}{Caption text}

\item{...}{
  Arguments passed on to \code{\link[=qto_src_span]{qto_src_span}}
  \describe{
    \item{\code{text}}{Caption or link text.}
  }}

\item{allow_missing}{If \code{FALSE} (default), error if string is not a URL and
not a valid web path. If \code{TRUE}, allow input to be a path to a non-existent
file.}

\item{allow_empty}{If \code{FALSE} (default), error if string is empty. If \code{TRUE},
allow input to be an empty string.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Simplified helper for qto_fig()
}
\keyword{internal}
