% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{qto_link}
\alias{qto_link}
\title{Create a link}
\usage{
qto_link(src, text, ..., allow_empty = FALSE, call = caller_env())
}
\arguments{
\item{src}{Image source or URL.}

\item{allow_empty}{If \code{FALSE} (default), error if string is empty. If \code{TRUE},
allow input to be an empty string.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Create a link
}
\seealso{
Other span: 
\code{\link{qto_fig}()},
\code{\link{qto_span}()}
}
\concept{span}
