% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{qto_list}
\alias{qto_list}
\alias{qto_ol}
\alias{qto_li}
\title{Create ordered or unordered lists for Quarto}
\usage{
qto_ol(.list = NULL, ..., level = 1)

qto_li(.list = NULL, ..., level = 1, symbol = "*")
}
\arguments{
\item{.list}{A vector of list items. Optional if items are passed to \code{...}.}

\item{...}{Items to use as list items.}

\item{level}{Indent level of list. Nested levels in a single function call
are not yet supported.}

\item{symbol}{Symbol to use for list bullet. Use "(@)" to create a list where
numbering continues after interruption.}
}
\description{
\code{\link[=qt_ol]{qt_ol()}} created ordered lists and \code{\link[=qt_li]{qt_li()}} creates unordered lists. See
the Quarto documentation for more information on lists:
\url{https://quarto.org/docs/authoring/markdown-basics.html#lists}
}
\examples{
qto_ol(LETTERS[1:3])

qto_li(LETTERS[4:6])

qto_li(LETTERS[7:9], level = 2)

}
\seealso{
\code{\link[=qto_dl]{qto_dl()}}
}
