% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcodes.R
\name{qto_shortcode}
\alias{qto_shortcode}
\alias{qto_pagebreak}
\alias{qto_video}
\alias{qto_kbd}
\title{Create a Quarto shortcode}
\usage{
qto_shortcode(x, before = "{{< ", after = " >}}", ...)

qto_video(src)

qto_pagebreak()

qto_kbd(...)
}
\description{
Create a Quarto shortcode

Videos \url{https://quarto.org/docs/authoring/markdown-basics.html#videos}

Page breaks \url{https://quarto.org/docs/authoring/markdown-basics.html#page-breaks}

Keyboard shortcuts \url{https://quarto.org/docs/authoring/markdown-basics.html#keyboard-shortcuts}
}
\examples{
qto_video("https://www.youtube.com/embed/wo9vZccmqwc")

qto_pagebreak()

qto_kbd("Shift", "Ctrl", "P")

qto_kbd(
  mac = "Shift-Command-O",
  win = "Shift-Control-O",
  linux = "Shift-Ctrl-L"
)

}
