% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.R
\name{qto_span}
\alias{qto_span}
\title{Create a Quarto span}
\usage{
qto_span(text, ..., .attributes = NULL)
}
\arguments{
\item{text}{Text to wrap in span bracket.}

\item{...}{
  Arguments passed on to \code{\link[=qto_attributes]{qto_attributes}}
  \describe{
    \item{\code{id}}{Div or span identifier}
    \item{\code{class}}{Div or span class}
    \item{\code{css}}{If \code{{htmltools}} is installed, a list of css style attributes to
pass to \code{\link[htmltools:css]{htmltools::css()}}.}
    \item{\code{.output}}{Ouput type. If "embrace", the returned attributes are enclosed
in curley brackets.}
    \item{\code{.drop_empty}}{If \code{TRUE}, empty attributes are dropped.}
  }}

\item{.attributes}{Optional list of attributes. If supplied, any attributes
passed to \code{...} are ignored.}
}
\description{
See the Quarto documentation for more on divs and spans:
\url{https://quarto.org/docs/authoring/markdown-basics.html#divs-and-spans}
}
\examples{

qto_span("This is *some text*", class = "class", key = "value")

qto_span("This is good", id = "id", class = "class", key1 = "val1", key2 = "val2")

}
