% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{qto_src_span}
\alias{qto_src_span}
\title{Simplified helper for \code{qto_link()}}
\usage{
qto_src_span(
  src,
  text = NULL,
  .before = "",
  ...,
  allow_missing = FALSE,
  allow_empty = FALSE,
  call = caller_env()
)
}
\arguments{
\item{src}{Image source or URL.}

\item{...}{
  Arguments passed on to \code{\link[=qto_attributes]{qto_attributes}}
  \describe{
    \item{\code{id}}{Div or span identifier. If \code{id} does not start with \code{"#"}, the
hash character is applied as a prefix.}
    \item{\code{class}}{Div or span class. If \code{class} does not start with \code{"."}, the
period character is applied as a prefix.}
    \item{\code{css}}{If \code{{htmltools}} is installed, a list of css style attributes to
pass to \code{\link[htmltools:css]{htmltools::css()}}.}
    \item{\code{.attributes}}{Optional list of attributes. If supplied, any attributes
passed to \code{...} are ignored.}
    \item{\code{.output}}{Output type. If "embrace", the returned attributes are
enclosed in curly brackets.}
    \item{\code{.drop_empty}}{If \code{TRUE}, empty attributes are dropped.}
  }}

\item{allow_missing}{If \code{FALSE} (default), error if string is not a URL and
not a valid web path. If \code{TRUE}, allow input to be a path to a non-existent
file.}

\item{allow_empty}{If \code{FALSE} (default), error if string is empty. If \code{TRUE},
allow input to be an empty string.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{caption}{Caption text}
}
\description{
Also used by \code{\link[=qto_fig_span]{qto_fig_span()}}
}
\keyword{internal}
