% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartools-package.R
\docType{package}
\name{quartools-package}
\alias{quartools}
\alias{quartools-package}
\title{quartools: programmatic element creation}
\description{
This package allows for the creation of quarto markdown via R function calls.
}
\section{todo}{
NA
}

\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

}
\keyword{internal}
