% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{span}
\alias{span}
\title{Create a quarto span element}
\usage{
span(..., attr = NULL)
}
\arguments{
\item{...}{todoc}

\item{attr}{todoc}
}
\value{
character vector of length 1
}
\description{
Create a span element with optional attributes to assign to it. Dots
is collapsed into a single character vector with a space between each element.

If attr is a vector with a length > 1,
attributes will be combined into a single character vector via \link[base:paste]{base::paste0}.
}
\examples{
span("Hello world!", attr = ".bold")
}
\seealso{
Other elements: 
\code{\link{div}()}
}
\concept{elements}
