% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_margin_column}
\alias{with_margin_column}
\title{TODO}
\usage{
with_margin_column(...)
}
\arguments{
\item{...}{todo}
}
\description{
TODO
}
\seealso{
Other page layout functions: 
\code{\link{with_body_column}()},
\code{\link{with_page_column}()},
\code{\link{with_screen_column}()},
\code{\link{with_screen_inset_column}()}
}
\concept{page layout functions}
