% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_screen_inset_column}
\alias{with_screen_inset_column}
\title{TODO}
\usage{
with_screen_inset_column(..., extension = NULL)
}
\arguments{
\item{...}{todo}

\item{extension}{todo}

\item{outset}{todo}
}
\description{
TODO
}
\seealso{
Other page layout functions: 
\code{\link{with_body_column}()},
\code{\link{with_margin_column}()},
\code{\link{with_page_column}()},
\code{\link{with_screen_column}()}
}
\concept{page layout functions}
