% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{as_markdown}
\alias{as_markdown}
\title{Convert to markdown chunk}
\usage{
as_markdown(x, collapse = "")
}
\arguments{
\item{x}{object to convert to character vector}

\item{collapse}{arg passed to \link[base:paste]{base::paste0}}
}
\value{
character vector of length 1
}
\description{
Given x, collapse each element of x so that a character vector
of length 1 is returned, and order knitr to render the result "asis".
}
