% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{div}
\alias{div}
\title{Create a quarto div element}
\usage{
div(..., attr = NULL)
}
\arguments{
\item{...}{unnamed values that can be coerced to character vectors}

\item{attr}{values that can be coerced to character vectors}
}
\value{
character vector of length 1
}
\description{
Create a div block element with optional attributes to assign to it. Dots
are collapsed into a single character vector with a newline between each
element.

If attr is a vector with a length > 1, attributes will be combined into
a single character vector via \link[base:paste]{base::paste0}.
}
\examples{
# Create a quarto tip callout
div("Hello world!", attr = ".callout-tip")

# div calls can be nested
div(
  div("Hello world!"),
  attr = ".my_class"
)
}
\seealso{
Other elements: 
\code{\link{span}()}
}
\concept{elements}
\keyword{internal}
