% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_body_column}
\alias{with_body_column}
\alias{with_page_column}
\alias{with_screen_inset_column}
\alias{with_screen_column}
\alias{with_margin_column}
\title{Temporarily modify page layout}
\usage{
with_body_column(..., outset = FALSE, extension = NULL)

with_page_column(..., extension = NULL)

with_screen_inset_column(..., extension = NULL)

with_screen_column(..., extension = NULL)

with_margin_column(...)
}
\arguments{
\item{...}{arguments to pass to \link{div}}

\item{outset}{should the outset affix be applied to the column class?}

\item{extension}{affix class to apply to column class}
}
\description{
Create a div block that modifies the current quarto layout column temporarily.

See \href{https://quarto.org/docs/authoring/article-layout.html#available-columns}{the Quarto documentation for a full list of available extension options}.
}
\examples{
with_body_column("Hello world!")
}
