% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{qto_block}
\alias{qto_block}
\title{Create a block of text to render as Markdown text in Quarto}
\usage{
qto_block(..., collapse = "", call = caller_env())
}
\arguments{
\item{...}{dots to convert to character vector}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.  When \code{collapse} is a string,
    the result is always a string (\code{\link[base]{character}} of length 1).}
}
\value{
character vector of length 1
}
\description{
\code{\link[=qto_block]{qto_block()}} based on \code{\link[=as_markdown]{as_markdown()}}
}
\examples{
as_markdown(c("Hello world!"))
}
\seealso{
\itemize{
\item \code{\link[=as_markdown]{as_markdown()}}
\item \code{\link[knitr:asis_output]{knitr::asis_output()}}
}
}
