% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callout.R
\name{qto_callout}
\alias{qto_callout}
\title{Create a Quarto callout block}
\usage{
qto_callout(
  ...,
  type = c("note", "tip", "warning", "caution", "important"),
  heading = NULL,
  collapse = NULL,
  appearance = NULL,
  icon = NULL,
  title = NULL,
  id = NULL,
  class = NULL,
  .attributes = NULL,
  level = 2,
  sep = " ",
  fence = ":::"
)
}
\arguments{
\item{...}{dots to convert to character vector}

\item{type}{Callout type. One of "note", "tip", "warning", "caution", or
"important". \url{https://quarto.org/docs/authoring/callouts.html#callout-types}}

\item{collapse}{If \code{TRUE}, create a folded callout. \url{https://quarto.org/docs/authoring/callouts.html#collapse}}

\item{appearance}{Callout appearance. One of "default", "simple", "minimal".
\url{https://quarto.org/docs/authoring/callouts.html#appearance}}

\item{icon}{If \code{FALSE}, create a callout with an icon. \url{https://quarto.org/docs/authoring/callouts.html#icons}}

\item{id}{Div or span identifier}

\item{class}{Div or span class}

\item{.attributes}{Optional list of attributes. If supplied, any attributes
passed to \code{...} are ignored.}

\item{fence}{If numeric, \code{fence} must be a minimum of 3 and sets the number
of times the standard fence character ":" should be repeated. If character,
\code{fence} is used as is.}
}
\description{
Callouts are an excellent way to draw extra attention to certain concepts, or
to more clearly indicate that certain content is supplemental or applicable
to only some scenarios. \url{https://quarto.org/docs/authoring/callouts.html}
}
\examples{

qto_callout(
"Callouts provide a simple way to attract attention, for example, to this warning.",
type = "warning"
)

qto_callout(
"This is an example of a callout with a title.",
type = "tip",
title = "Tip with a title"
)

}
