% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{qto_div}
\alias{qto_div}
\title{Create a Quarto div with optional classes, attributes, and other identifiers}
\usage{
qto_div(
  ...,
  id = NULL,
  class = NULL,
  css = NULL,
  .attributes = NULL,
  .content = NULL,
  collapse = "",
  fence = ":::",
  drop_empty = TRUE,
  drop_na = TRUE,
  call = caller_env()
)
}
\arguments{
\item{...}{dots to convert to character vector}

\item{id}{Div or span identifier}

\item{class}{Div or span class}

\item{css}{If \code{{htmltools}} is installed, a list of css style attributes to
pass to \code{\link[htmltools:css]{htmltools::css()}}.}

\item{.attributes}{Optional list of attributes. If supplied, any attributes
passed to \code{...} are ignored.}

\item{.content}{If \code{.content} is supplied, any values passed to \code{...} are
ignored. If \code{.content} is \code{NULL}, it is set as all values passed to \code{...}.}

\item{collapse}{Passed to \code{\link[base:paste]{base::paste0()}} with \code{.content.}}

\item{fence}{If numeric, \code{fence} must be a minimum of 3 and sets the number
of times the standard fence character ":" should be repeated. If character,
\code{fence} is used as is.}

\item{drop_empty}{If \code{TRUE}, drop empty values from \code{.content} or \code{...}}

\item{drop_na}{If \code{TRUE}, drop \code{NA} values from \code{.content} or \code{...}}
}
\description{
\url{https://quarto.org/docs/authoring/markdown-basics.html#divs-and-spans}
}
\examples{

# div with an ID
qto_div(
  "This content can be styled with a border",
  class = "border"
)

# Nested div
qto_div(
  qto_div("Here is a warning.", class = "warning"),
  "More content.",
  id = "special",
  class = "sidebar",
  fence = 5
)

}
\seealso{
\itemize{
\item \code{\link[=div]{div()}}
\item \code{\link[=qto_callout_block]{qto_callout_block()}}
}
}
