% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{qto_fig}
\alias{qto_fig}
\title{Insert an figure or image}
\usage{
qto_fig(
  src,
  caption = "",
  alt = NULL,
  reference = NULL,
  align = NULL,
  pos = NULL,
  width = NULL,
  height = NULL,
  title = NULL,
  ...,
  allow_missing = TRUE
)
}
\arguments{
\item{allow_missing}{If \code{FALSE}, error if src is not an existing file or a
valid URL. URLs are not checked if they work.}
}
\description{
\url{https://quarto.org/docs/authoring/figures.html}
}
\examples{
qto_img("image.jpeg", "Image caption", alt = "Image alt text")

qto_img("graphic.pdf", title = "Title of the graphic")

}
