% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{qto_list}
\alias{qto_list}
\alias{qto_ol}
\alias{qto_li}
\title{Create ordered or unordered lists for Quarto}
\usage{
qto_ol(.list = NULL, ..., level = 1, symbol = "1.")

qto_li(.list = NULL, ..., level = 1, symbol = "*", sep = "")
}
\arguments{
\item{.list}{A vector of list items. Optional if items are passed to \code{...}.}

\item{...}{Items to use as list items. Ignored if .list is supplied.}

\item{level}{Indent level of list. Nested levels in a single function call
are not yet supported.}

\item{symbol}{Symbol to use for list bullet. Use "(@)" to create a list where
numbering continues after interruption.}

\item{sep}{Separator character between list items passed to \code{\link[=qto_block]{qto_block()}}.
Defaults to "". Set to \code{"\\n"} for wider list spacing.}
}
\description{
\code{\link[=qto_ol]{qto_ol()}} created ordered lists and \code{\link[=qto_li]{qto_li()}} creates unordered lists. See
the Quarto documentation for more information on lists:
\url{https://quarto.org/docs/authoring/markdown-basics.html#lists}
}
\examples{
qto_ol(LETTERS[1:2])

qto_li(LETTERS[3:4])

qto_li(LETTERS[5:6], level = 2)

qto_ol(LETTERS[7:8], sep = "\n")

qto_li(LETTERS[5:6], symbol = c("*", "    +"))

}
\seealso{
\code{\link[=qto_dl]{qto_dl()}}
}
