% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcodes.R
\name{qto_shortcode}
\alias{qto_shortcode}
\alias{qto_pagebreak}
\alias{qto_video}
\alias{qto_kbd}
\title{Create a Quarto shortcode}
\usage{
qto_shortcode(x, before = "{{< ", after = " >}}", ...)

qto_video(src)

qto_pagebreak()

qto_kbd(...)
}
\arguments{
\item{x}{An input string.}

\item{before, after}{Text to insert before and after the supplied string.}

\item{...}{For \code{\link[=qto_kbd]{qto_kbd()}}, a set of strings to combine with "-" or a named
set of attributes where names are one of "mac", "win", or "linux".}

\item{src}{For \code{\link[=qto_video]{qto_video()}}, a URL with the video source.}
}
\description{
\code{\link[=qto_shortcode]{qto_shortcode()}} creates a shortcode based on an input string.
}
\examples{
qto_video("https://www.youtube.com/embed/wo9vZccmqwc")

qto_pagebreak()

qto_kbd("Shift", "Ctrl", "P")

qto_kbd(
  mac = "Shift-Command-O",
  win = "Shift-Control-O",
  linux = "Shift-Ctrl-L"
)

}
