% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartools-package.R
\docType{package}
\name{quartools-package}
\alias{quartools}
\alias{quartools-package}
\title{quartools: Programmatic Element Creation For Quarto Documents}
\description{
Programatically generate quarto-compliant markdown elements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElianHugh/quartools}
  \item Report bugs at \url{https://github.com/ElianHugh/quartools/issues}
}

}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID})
}

}
\keyword{internal}
