% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartools-package.R
\docType{package}
\name{quartools-package}
\alias{quartools}
\alias{quartools-package}
\title{quartools: Programmatic Element Creation For Quarto Documents}
\description{
Generate quarto-compliant markdown elements programatically.
}
\author{
\strong{Maintainer}: Elian Thiele-Evans \email{elianhte@gmail.com} (\href{https://orcid.org/0000-0001-8008-3165}{ORCID})

}
\keyword{internal}
